Refactoring and renaming, fix imports
