+++ CustomAlteryxToolsNet/CustomAlteryxTools.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 15.0.27428.2037
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CustomAlteryxTools", "CustomAlteryxTools.csproj", "{44182C44-3979-48EE-9D25-E0D86E48F4EB}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SandboxAndTests", "..\SandboxAndTests\SandboxAndTests.csproj", "{D2502DFF-332D-4B14-9F12-DB7C3C6E574C}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -29,18 +27,6 @@ Global
 		{44182C44-3979-48EE-9D25-E0D86E48F4EB}.Release|x64.Build.0 = Release|x64
 		{44182C44-3979-48EE-9D25-E0D86E48F4EB}.Release|x86.ActiveCfg = Release|Any CPU
 		{44182C44-3979-48EE-9D25-E0D86E48F4EB}.Release|x86.Build.0 = Release|Any CPU
-		{D2502DFF-332D-4B14-9F12-DB7C3C6E574C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D2502DFF-332D-4B14-9F12-DB7C3C6E574C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D2502DFF-332D-4B14-9F12-DB7C3C6E574C}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{D2502DFF-332D-4B14-9F12-DB7C3C6E574C}.Debug|x64.Build.0 = Debug|Any CPU
-		{D2502DFF-332D-4B14-9F12-DB7C3C6E574C}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{D2502DFF-332D-4B14-9F12-DB7C3C6E574C}.Debug|x86.Build.0 = Debug|Any CPU
-		{D2502DFF-332D-4B14-9F12-DB7C3C6E574C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D2502DFF-332D-4B14-9F12-DB7C3C6E574C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D2502DFF-332D-4B14-9F12-DB7C3C6E574C}.Release|x64.ActiveCfg = Release|Any CPU
-		{D2502DFF-332D-4B14-9F12-DB7C3C6E574C}.Release|x64.Build.0 = Release|Any CPU
-		{D2502DFF-332D-4B14-9F12-DB7C3C6E574C}.Release|x86.ActiveCfg = Release|Any CPU
-		{D2502DFF-332D-4B14-9F12-DB7C3C6E574C}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
