+++ GlassdoorScrape/GlassdoorScrape/GlassdoorScrapeCore/GlassdoorScrape_Test.py
@@ -1,5 +1,6 @@
 import unittest
-import GlassdoorScrapeCore.GlassdoorScrapeCore as Core
+import GlassdoorScrapeCompany as Core
+import GlassdoorScrapeUtilities as Ut
 
 
 class MyTestCase(unittest.TestCase):
@@ -25,6 +26,18 @@ class MyTestCase(unittest.TestCase):
         self.assertEqual(10, len(result.reviews))
         self.assertEqual('E351220', result.GlassdoorId)
         self.assertEqual(10, len(result.interviews))
+
+    def test_StripHtml(self):
+        test_strings = [
+            "<div>Something</div>",
+            "<div property='blah' >Something</div>",
+            "<div><span><br>Something</span></div>"
+        ]
+
+        for test_str in test_strings:
+            clean_str = Ut.strip_html(test_str)
+            print(clean_str)
+            self.assertEqual("Something", clean_str)
 
 
 if __name__ == '__main__':
